fixed bug: refreshing frequently caused logout automatically(every refresh reload tokens to when checking local storage it gave empty token, that is why auto logout bug happened)

+++ auctionhub-Frontend/context/AuthContext.jsx
@@ -30,9 +30,15 @@ export const AuthProvider = ({ children }) => {
         } catch (error) {
             console.log(error);
             // logout();
-            setUser(null);
-            setIsAuthenticated(false);
-            localStorage.clear();
+            // setUser(null);
+            // setIsAuthenticated(false);
+            // localStorage.clear();
+            const refreshToken = localStorage.getItem("refresh");
+            if (!refreshToken) {
+                setUser(null);
+                setIsAuthenticated(false);
+                localStorage.clear();
+            }
         } finally {
             setLoading(false);
         }
